Split out WriteError from Error, bump MSRV to 1.85 and edition to 2024, removed pinned deps
