+++ .github/workflows/rosbag.yml
@@ -20,24 +20,10 @@ jobs:
       - uses: actions/checkout@v1
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.63.0 # MSRV
+          toolchain: 1.85.0 # MSRV
           components: clippy
           override: true
           profile: minimal
-      # Pin dependencies for MSRV compatibility (Rust 1.63)
-      # Order matters: pin high-level crates before their transitive dependencies
-      - name: Pin MSRV-compatible dependency versions
-        run: |
-          cargo update -p log --precise 0.4.21
-          cargo update -p bzip2-sys --precise 0.1.11+1.0.8
-          cargo update -p tempfile --precise 3.10.1
-          cargo update -p backtrace --precise 0.3.67
-          cargo update -p serde --precise 1.0.197
-          cargo update -p serde_derive --precise 1.0.197
-          cargo update -p syn --precise 2.0.52
-          cargo update -p quote --precise 1.0.35
-          cargo update -p proc-macro2 --precise 1.0.78
-          cargo update -p unicode-ident --precise 1.0.12
       - run: cargo clippy --all -- -D warnings
 
   rustfmt:
@@ -79,7 +65,7 @@ jobs:
     strategy:
       matrix:
         rust:
-          - 1.63.0 # MSRV
+          - 1.85.0 # MSRV
           - stable
     steps:
       - uses: actions/checkout@v1
@@ -88,21 +74,6 @@ jobs:
           profile: minimal
           toolchain: ${{ matrix.rust }}
           override: true
-      # Pin dependencies for MSRV compatibility (Rust 1.63)
-      # Order matters: pin high-level crates before their transitive dependencies
-      - name: Pin MSRV-compatible dependency versions
-        if: matrix.rust == '1.63.0'
-        run: |
-          cargo update -p log --precise 0.4.21
-          cargo update -p bzip2-sys --precise 0.1.11+1.0.8
-          cargo update -p tempfile --precise 3.10.1
-          cargo update -p backtrace --precise 0.3.67
-          cargo update -p serde --precise 1.0.197
-          cargo update -p serde_derive --precise 1.0.197
-          cargo update -p syn --precise 2.0.52
-          cargo update -p quote --precise 1.0.35
-          cargo update -p proc-macro2 --precise 1.0.78
-          cargo update -p unicode-ident --precise 1.0.12
       - run: cargo test
       - run: cargo test --all-features
