+++ Cargo.toml
@@ -4,8 +4,8 @@ version = "0.6.3"
 description = "Utilities for reading ROS bag files."
 authors = ["Artyom Pavlov <newpavlov@gmail.com>"]
 license = "MIT OR Apache-2.0"
-edition = "2021"
-rust-version = "1.63"
+edition = "2024"
+rust-version = "1.85"
 readme = "README.md"
 documentation = "https://docs.rs/rosbag"
 repository = "https://github.com/newpavlov/rosbag-rs"
